requirements.txt
fastapi
uvicorn
python-dotenv
SQLAlchemy
SQLAlchemy-Utils
PyJWT
pydantic
pydantic[email]
bcrypt
